Fix vscode workspace switching
Using the URL is much faster than the `code` executable, but using the
URL directly on a file cause the file to open in whatever vscode
workspace is considered active. Since it's a URL, there's no obvious
way to "set the PWD" for the opening process. However, issuing `open`
twice in succession seems to work well (first for the project dir,
then for the file+line). E.g.

open 'cursor://file//Users/dan/src/delta'
open 'cursor://file//Users/dan/src/delta/src/main.rs:7'

However, this only works if the workspace is open already; otherwise
it kills the current workspace and starts the new one. Therefore we
open the editor workspace (via the command line executable) if the
project is not open already. We do this synchronously, so that it is
done by the time the terminal and editor threads run.

Also, do not open directory with a line number as if it were a file